Updated Allowances  and Deductions
Updated how we get and calculate allowances and deductions.

Now uses options to grab both separately and you can have both.

Works out the calculation for the user and adds/removes it from the tax free amount

Displays result as a modification to tax free amount rather than as a separate line item on the results screens.