Added logic for tapering and taxcode messages
When a user uses a salary above 100K and has a default tax code shows a message on check answers.

When a user has used a specific tax code other than the default and then changes the salary show a message to highlight that they might need to check it.

Added flags to turn on/off messages when an update has been made.