Ignore tapering when user uses 1257L over £100K
Previously the prototype would taper over £100K regardless of if the user used a tax code. In order to test messaging on the results screen should a user ignore the initial message on check answers about 1257L over £100K the prototype was updated to accept the tax code and display a second warning on check answers.